mide/DICOM_doc_helper/mide/requirements.txt
numpy==1.20.3
opencv-python==4.5.2.54
packaging==20.9
pandas==1.3.0
Pillow==8.2.0
psutil==5.8.0
pydicom==2.1.2
pynetdicom==1.5.7
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
schedule==1.1.0
toml==0.10.2
virtualenv==20.4.7
pytesseract==0.3.7
GDCM == 1.1
pylibjpeg == 1.3.0
pylibjpeg-libjpeg == 1.2